+++ FE/src/components/common/icons/Email.tsx
