+++ FE/src/components/common/icons/ISchool.tsx
