+++ FE/src/components/common/icons/Logo.tsx
