+++ FE/src/components/common/icons/Password.tsx
