+++ FE/src/components/common/icons/Toggle.tsx
