+++ FE/src/components/join/AgreementCheckbox.tsx
