+++ FE/src/components/join/DuplicateCheckButton.tsx
