+++ FE/src/components/join/JoinInput.tsx
