+++ FE/src/components/join/SubmitBtn.tsx
@@ -1,5 +1,5 @@
 "use client";
-import StyledBtn from "./ui/StyledBtn";
+import StyledBtn from "../common/StyledBtn";
 
 export default function SubmitBtn() {
   return <StyledBtn text="회원가입" />;
