+++ FE/src/components/join/ToggleBtn.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { Toggle } from "./ui/icon";
+import { Toggle } from "../common/icon";
 
 export default function ToggleBtn() {
   return <Toggle />;
